radiation puddle added
does heavy damage to the player

Added the radiation puddles that are supposed to be in the radiator springs for Liam to add a platforming level as per plans. The puddles deal a lot of damage but the box collider causing the damage is very small, so it’ll only damage the player if they fall into it.